feat(inference): add converting to onnx format for inference
