+++ pyproject.toml
@@ -20,6 +20,8 @@ tqdm = "^4.66.1"
 numpy = "^1.26.2"
 pillow = "^10.1.0"
 torchmetrics = {extras = ["detection"], version = "^1.2.0"}
+onnx = "^1.15.0"
+hydra-core = "^1.3.2"
 
 
 [tool.poetry.group.lint.dependencies]
