--- sign_recognition/visualization/evaluate.py
@@ -1,1 +1,0 @@
-# todo here will be the class that will be used to evaluate the model
